Moving the storageClasses for the default.
